feat(new source)!: Merge existing 'tcp' and 'udp' sources into a single 'socket' source (#1211)
