+++ .meta/sources/socket.toml
@@ -1,23 +1,34 @@
-[sources.tcp]
-title = "TCP"
+[sources.socket]
+title = "Socket"
 common = true
 delivery_guarantee = "best_effort"
 function_category = "receive"
 guides = []
 output_types = ["log"]
 resources = []
-through_description = "the TCP protocol"
+through_description = "a socket, such as a TCP, UDP, or Unix socket"
 
-[sources.tcp.options.address]
+[sources.socket.options.mode]
+type = "string"
+common = true
+null = false
+description = "The type of socket to use."
+
+[sources.socket.options.mode.enum]
+tcp = "The TCP protocol."
+udp = "The UDP protocol."
+
+[sources.socket.options.address]
 type = "string"
 common = true
 examples = ["0.0.0.0:9000", "systemd", "systemd#3"]
 null = false
 description = """The address to listen for connections on, \
-or "systemd#N" to use the Nth socket passed by systemd socket activation.\
+or "systemd#N" to use the Nth socket passed by systemd socket activation. \
+Valid for "tcp" and "udp" sockets.\
 """
 
-[sources.tcp.options.host_key]
+[sources.socket.options.host_key]
 name = "host_key"
 type = "string"
 category = "Context"
@@ -25,48 +36,44 @@ null = false
 default = "host"
 section = "context"
 description = """\
-The key name added to each event representing the current host.\
+The key name added to each event representing the current host. \
+Valid for "tcp" and "udp" sockets.\
 """
 
-[sources.tcp.options.max_length]
+[sources.socket.options.max_length]
 type = "int"
 default = 102400
 null = true
 unit = "bytes"
 description = """\
-The maximum bytes size of incoming messages before they are discarded.\
+The maximum bytes size of incoming messages before they are discarded. \
+Valid for "tcp" and "udp" sockets.\
 """
 
-[sources.tcp.options.shutdown_timeout_secs]
+[sources.socket.options.shutdown_timeout_secs]
 type = "int"
 default = 30
 null = false
 unit = "seconds"
 description = """\
-The timeout before a connection is forcefully closed during shutdown.\
+The timeout before a connection is forcefully closed during shutdown. \
+Valid only for "tcp" sockets.\
 """
 
-[[sources.tcp.output.log.examples]]
+[[sources.socket.output.log.examples]]
 label = "Generic"
-body = """\
-Given the following input:
-
-```text
-2019-02-13T19:48:34+00:00 [info] Started GET "/" for 127.0.0.1
-```
-
-A log event will be output with the following structure:
-
-```json
+input = """\
+2019-02-13T19:48:34+00:00 [info] Started GET "/" for 127.0.0.1\
+"""
+output = """\
 {
   "timestamp": <current_timestamp>,
   "message": "2019-02-13T19:48:34+00:00 [info] Started GET "/" for 127.0.0.1",
   "host": "<upstream_hostname>"
 }
-```\
 """
 
-[sources.tcp.output.log.fields.host]
+[sources.socket.output.log.fields.host]
 type = "string"
 examples = ["my.host.com"]
 optional = false
@@ -74,7 +81,7 @@ description = """\
 The upstream hostname.
 """
 
-[sources.tcp.output.log.fields.message]
+[sources.socket.output.log.fields.message]
 type = "string"
 examples = ["Started GET / for 127.0.0.1 at 2012-03-10 14:28:14 +0100"]
 optional = false
@@ -82,7 +89,7 @@ description = """\
 The raw message, unaltered.
 """
 
-[sources.tcp.output.log.fields.timestamp]
+[sources.socket.output.log.fields.timestamp]
 type = "timestamp"
 examples = ["2019-11-01T21:15:47.443232Z"]
 optional = false
