--- .meta/sources/udp.toml
@@ -1,80 +1,0 @@
-[sources.udp]
-title = "UDP"
-common = false
-delivery_guarantee = "best_effort"
-function_category = "receive"
-guides = []
-output_types = ["log"]
-resources = []
-through_description = "the UDP protocol"
-
-[sources.udp.options.address]
-type = "string"
-common = true
-examples = ["0.0.0.0:9000"]
-null = false
-description = "The address to bind the socket to."
-
-[sources.udp.options.host_key]
-name = "host_key"
-type = "string"
-category = "Context"
-null = false
-default = "host"
-section = "context"
-description = """\
-The key name added to each event representing the current host.\
-"""
-
-[sources.udp.options.max_length]
-type = "int"
-default = 102400
-null = true
-unit = "bytes"
-description = """\
-The maximum bytes size of incoming messages before they are discarded.\
-"""
-
-[[sources.udp.output.log.examples]]
-label = "Generic"
-body = """\
-Given the following input:
-
-```text
-2019-02-13T19:48:34+00:00 [info] Started GET "/" for 127.0.0.1
-```
-
-A log event will be output with the following structure:
-
-```json
-{
-  "timestamp": <current_timestamp>,
-  "message": "2019-02-13T19:48:34+00:00 [info] Started GET "/" for 127.0.0.1",
-  "host": "<upstream_hostname>"
-}
-```\
-"""
-
-[sources.udp.output.log.fields.host]
-type = "string"
-examples = ["my.host.com"]
-optional = false
-description = """\
-The upstream hostname.
-"""
-
-[sources.udp.output.log.fields.message]
-type = "string"
-examples = ["Started GET / for 127.0.0.1 at 2012-03-10 14:28:14 +0100"]
-optional = false
-description = """\
-The raw message, unaltered.
-"""
-
-[sources.udp.output.log.fields.timestamp]
-type = "timestamp"
-examples = ["2019-11-01T21:15:47.443232Z"]
-optional = false
-description = """\
-The exact time the event was ingested.\
-"""
