+++ benches/http.rs
@@ -22,7 +22,10 @@ fn benchmark_http_no_compression(c: &mut Criterion) {
         b.iter_with_setup(
             || {
                 let mut config = config::Config::empty();
-                config.add_source("in", sources::tcp::TcpConfig::new(in_addr.into()));
+                config.add_source(
+                    "in",
+                    sources::socket::SocketConfig::make_tcp_config(in_addr),
+                );
                 config.add_sink(
                     "out",
                     &["in"],
@@ -70,7 +73,10 @@ fn benchmark_http_gzip(c: &mut Criterion) {
         b.iter_with_setup(
             || {
                 let mut config = config::Config::empty();
-                config.add_source("in", sources::tcp::TcpConfig::new(in_addr.into()));
+                config.add_source(
+                    "in",
+                    sources::socket::SocketConfig::make_tcp_config(in_addr),
+                );
                 config.add_sink(
                     "out",
                     &["in"],
