+++ src/sources/mod.rs
@@ -6,12 +6,11 @@ pub mod journald;
 #[cfg(feature = "rdkafka")]
 pub mod kafka;
 pub mod kubernetes;
+pub mod socket;
 pub mod splunk_hec;
 pub mod statsd;
 pub mod stdin;
 pub mod syslog;
-pub mod tcp;
-pub mod udp;
 mod util;
 pub mod vector;
