+++ tests/buffering.rs
@@ -23,7 +23,10 @@ fn test_buffering() {
 
     // Run vector while sink server is not running, and then shut it down abruptly
     let mut config = config::Config::empty();
-    config.add_source("in", sources::tcp::TcpConfig::new(in_addr.into()));
+    config.add_source(
+        "in",
+        sources::socket::SocketConfig::make_tcp_config(in_addr),
+    );
     config.add_sink(
         "out",
         &["in"],
@@ -51,7 +54,10 @@ fn test_buffering() {
 
     // Start sink server, then run vector again. It should send all of the lines from the first run.
     let mut config = config::Config::empty();
-    config.add_source("in", sources::tcp::TcpConfig::new(in_addr.into()));
+    config.add_source(
+        "in",
+        sources::socket::SocketConfig::make_tcp_config(in_addr),
+    );
     config.add_sink(
         "out",
         &["in"],
@@ -113,7 +119,10 @@ fn test_max_size() {
 
     // Run vector while sink server is not running, and then shut it down abruptly
     let mut config = config::Config::empty();
-    config.add_source("in", sources::tcp::TcpConfig::new(in_addr.into()));
+    config.add_source(
+        "in",
+        sources::socket::SocketConfig::make_tcp_config(in_addr),
+    );
     config.add_sink(
         "out",
         &["in"],
@@ -140,7 +149,10 @@ fn test_max_size() {
 
     // Start sink server, then run vector again. It should send the lines from the first run that fit in the limited space
     let mut config = config::Config::empty();
-    config.add_source("in", sources::tcp::TcpConfig::new(in_addr.into()));
+    config.add_source(
+        "in",
+        sources::socket::SocketConfig::make_tcp_config(in_addr),
+    );
     config.add_sink(
         "out",
         &["in"],
@@ -183,8 +195,14 @@ fn test_max_size_resume() {
     let out_addr = next_addr();
 
     let mut config = config::Config::empty();
-    config.add_source("in1", sources::tcp::TcpConfig::new(in_addr1.into()));
-    config.add_source("in2", sources::tcp::TcpConfig::new(in_addr2.into()));
+    config.add_source(
+        "in1",
+        sources::socket::SocketConfig::make_tcp_config(in_addr1),
+    );
+    config.add_source(
+        "in2",
+        sources::socket::SocketConfig::make_tcp_config(in_addr2),
+    );
     config.add_sink(
         "out",
         &["in1", "in2"],
@@ -237,7 +255,10 @@ fn test_reclaim_disk_space() {
 
     // Run vector while sink server is not running, and then shut it down abruptly
     let mut config = config::Config::empty();
-    config.add_source("in", sources::tcp::TcpConfig::new(in_addr.into()));
+    config.add_source(
+        "in",
+        sources::socket::SocketConfig::make_tcp_config(in_addr),
+    );
     config.add_sink(
         "out",
         &["in"],
@@ -274,7 +295,10 @@ fn test_reclaim_disk_space() {
 
     // Start sink server, then run vector again. It should send all of the lines from the first run.
     let mut config = config::Config::empty();
-    config.add_source("in", sources::tcp::TcpConfig::new(in_addr.into()));
+    config.add_source(
+        "in",
+        sources::socket::SocketConfig::make_tcp_config(in_addr),
+    );
     config.add_sink(
         "out",
         &["in"],
