+++ tests/config.rs
@@ -12,7 +12,8 @@ fn happy_path() {
     load(
         r#"
         [sources.in]
-        type = "tcp"
+        type = "socket"
+        mode = "tcp"
         address = "127.0.0.1:1235"
 
         [transforms.sampler]
@@ -33,7 +34,7 @@ fn happy_path() {
     load(
         r#"
         [sources]
-        in = {type = "tcp", address = "127.0.0.1:1235"}
+        in = {type = "socket", mode = "tcp", address = "127.0.0.1:1235"}
 
         [transforms]
         sampler = {type = "sampler", inputs = ["in"], rate = 10, pass_list = ["error"]}
@@ -67,7 +68,26 @@ fn missing_key() {
     let err = load(
         r#"
         [sources.in]
+        type = "socket"
+
+        [sinks.out]
         type = "tcp"
+        inputs = ["in"]
+        address = "127.0.0.1:9999"
+      "#,
+    )
+    .unwrap_err();
+
+    assert_eq!(err, vec!["missing field `mode` for key `sources.in`"]);
+}
+
+#[test]
+fn missing_key2() {
+    let err = load(
+        r#"
+        [sources.in]
+        type = "socket"
+        mode = "tcp"
 
         [sinks.out]
         type = "tcp"
@@ -85,7 +105,8 @@ fn bad_type() {
     let err = load(
         r#"
         [sources.in]
-        type = "tcp"
+        type = "socket"
+        mode = "tcp"
         address = "127.0.0.1:1234"
 
         [sinks.out]
@@ -105,7 +126,8 @@ fn nonexistant_input() {
     let err = load(
         r#"
         [sources.in]
-        type = "tcp"
+        type = "socket"
+        mode = "tcp"
         address = "127.0.0.1:1235"
 
         [transforms.sampler]
@@ -137,7 +159,8 @@ fn bad_regex() {
     let err = load(
         r#"
         [sources.in]
-        type = "tcp"
+        type = "socket"
+        mode = "tcp"
         address = "127.0.0.1:1235"
 
         [transforms.sampler]
@@ -161,7 +184,8 @@ fn bad_regex() {
     let err = load(
         r#"
         [sources.in]
-        type = "tcp"
+        type = "socket"
+        mode = "tcp"
         address = "127.0.0.1:1235"
 
         [transforms.parser]
@@ -187,7 +211,8 @@ fn good_regex_parser() {
     let result = load(
         r#"
         [sources.in]
-        type = "tcp"
+        type = "socket"
+        mode = "tcp"
         address = "127.0.0.1:1235"
 
         [transforms.parser]
@@ -214,7 +239,8 @@ fn good_tokenizer() {
     let result = load(
         r#"
         [sources.in]
-        type = "tcp"
+        type = "socket"
+        mode = "tcp"
         address = "127.0.0.1:1235"
 
         [transforms.parser]
@@ -242,7 +268,8 @@ fn bad_s3_region() {
     let err = load(
         r#"
         [sources.in]
-        type = "tcp"
+        type = "socket"
+        mode = "tcp"
         address = "127.0.0.1:1235"
 
         [sinks.out1]
@@ -304,11 +331,13 @@ fn warnings() {
     let warnings = load(
         r#"
         [sources.in1]
-        type = "tcp"
+        type = "socket"
+        mode = "tcp"
         address = "127.0.0.1:1235"
 
         [sources.in2]
-        type = "tcp"
+        type = "socket"
+        mode = "tcp"
         address = "127.0.0.1:1236"
 
         [transforms.sampler1]
@@ -346,7 +375,8 @@ fn cycle() {
     let errors = load(
         r#"
         [sources.in]
-        type = "tcp"
+        type = "socket"
+        mode = "tcp"
         address = "127.0.0.1:1235"
 
         [transforms.one]
@@ -390,7 +420,8 @@ fn disabled_healthcheck() {
     load(
         r#"
       [sources.in]
-      type = "tcp"
+      type = "socket"
+      mode = "tcp"
       address = "127.0.0.1:1234"
 
       [sinks.out]
