+++ tests/tcp.rs
@@ -17,7 +17,10 @@ fn pipe() {
     let out_addr = next_addr();
 
     let mut config = config::Config::empty();
-    config.add_source("in", sources::tcp::TcpConfig::new(in_addr.into()));
+    config.add_source(
+        "in",
+        sources::socket::SocketConfig::make_tcp_config(in_addr),
+    );
     config.add_sink(
         "out",
         &["in"],
@@ -53,7 +56,10 @@ fn sample() {
     let out_addr = next_addr();
 
     let mut config = config::Config::empty();
-    config.add_source("in", sources::tcp::TcpConfig::new(in_addr.into()));
+    config.add_source(
+        "in",
+        sources::socket::SocketConfig::make_tcp_config(in_addr),
+    );
     config.add_transform(
         "sampler",
         &["in"],
@@ -110,8 +116,14 @@ fn merge() {
     let out_addr = next_addr();
 
     let mut config = config::Config::empty();
-    config.add_source("in1", sources::tcp::TcpConfig::new(in_addr1.into()));
-    config.add_source("in2", sources::tcp::TcpConfig::new(in_addr2.into()));
+    config.add_source(
+        "in1",
+        sources::socket::SocketConfig::make_tcp_config(in_addr1),
+    );
+    config.add_source(
+        "in2",
+        sources::socket::SocketConfig::make_tcp_config(in_addr2),
+    );
     config.add_sink(
         "out",
         &["in1", "in2"],
@@ -168,7 +180,10 @@ fn fork() {
     let out_addr2 = next_addr();
 
     let mut config = config::Config::empty();
-    config.add_source("in", sources::tcp::TcpConfig::new(in_addr.into()));
+    config.add_source(
+        "in",
+        sources::socket::SocketConfig::make_tcp_config(in_addr),
+    );
     config.add_sink(
         "out1",
         &["in"],
@@ -217,8 +232,14 @@ fn merge_and_fork() {
     // out1 receives both in1 and in2
     // out2 receives in2 only
     let mut config = config::Config::empty();
-    config.add_source("in1", sources::tcp::TcpConfig::new(in_addr1.into()));
-    config.add_source("in2", sources::tcp::TcpConfig::new(in_addr2.into()));
+    config.add_source(
+        "in1",
+        sources::socket::SocketConfig::make_tcp_config(in_addr1),
+    );
+    config.add_source(
+        "in2",
+        sources::socket::SocketConfig::make_tcp_config(in_addr2),
+    );
     config.add_sink(
         "out1",
         &["in1", "in2"],
@@ -283,7 +304,10 @@ fn reconnect() {
     let out_addr = next_addr();
 
     let mut config = config::Config::empty();
-    config.add_source("in", sources::tcp::TcpConfig::new(in_addr.into()));
+    config.add_source(
+        "in",
+        sources::socket::SocketConfig::make_tcp_config(in_addr),
+    );
     config.add_sink(
         "out",
         &["in"],
