+++ website/docs/reference/sources/socket.md
@@ -1,0 +1,292 @@
+---
+delivery_guarantee: "best_effort"
+description: "The Vector `socket` source ingests data through a socket, such as a TCP, UDP, or Unix socket and outputs `log` events."
+event_types: ["log"]
+issues_url: https://github.com/timberio/vector/issues?q=is%3Aopen+is%3Aissue+label%3A%22source%3A+socket%22
+operating_systems: ["linux","macos","windows"]
+sidebar_label: "socket|[\"log\"]"
+source_url: https://github.com/timberio/vector/tree/master/src/sources/socket.rs
+status: "prod-ready"
+title: "Socket Source"
+unsupported_operating_systems: []
+---
+
+The Vector `socket` source ingests data through a socket, such as a TCP, UDP, or Unix socket and outputs [`log`][docs.data-model.log] events.
+
+<!--
+     THIS FILE IS AUTOGENERATED!
+
+     To make changes please edit the template located at:
+
+     website/docs/reference/sources/socket.md.erb
+-->
+
+## Configuration
+
+import Tabs from '@theme/Tabs';
+
+<Tabs
+  block={true}
+  defaultValue="common"
+  values={[
+    { label: 'Common', value: 'common', },
+    { label: 'Advanced', value: 'advanced', },
+  ]
+}>
+
+import TabItem from '@theme/TabItem';
+
+<TabItem value="common">
+
+import CodeHeader from '@site/src/components/CodeHeader';
+
+<CodeHeader fileName="vector.toml" learnMoreUrl="/docs/setup/configuration/"/ >
+
+```toml
+[sources.my_source_id]
+  type = "socket" # must be: "socket"
+  address = "0.0.0.0:9000" # example
+  mode = "tcp" # example, enum
+```
+
+</TabItem>
+<TabItem value="advanced">
+
+<CodeHeader fileName="vector.toml" learnMoreUrl="/docs/setup/configuration/" />
+
+```toml
+[sources.my_source_id]
+  # REQUIRED - General
+  type = "socket" # must be: "socket"
+  address = "0.0.0.0:9000" # example
+  mode = "tcp" # example, enum
+
+  # OPTIONAL - General
+  max_length = 102400 # default, bytes
+  shutdown_timeout_secs = 30 # default, seconds
+
+  # OPTIONAL - Context
+  host_key = "host" # default
+```
+
+</TabItem>
+
+</Tabs>
+
+## Options
+
+import Fields from '@site/src/components/Fields';
+
+import Field from '@site/src/components/Field';
+
+<Fields filters={true}>
+
+
+<Field
+  common={true}
+  defaultValue={null}
+  enumValues={null}
+  examples={["0.0.0.0:9000","systemd","systemd#3"]}
+  name={"address"}
+  nullable={false}
+  path={null}
+  relevantWhen={null}
+  required={true}
+  templateable={false}
+  type={"string"}
+  unit={null}
+  >
+
+### address
+
+The address to listen for connections on, or "systemd#N" to use the Nth socket passed by systemd socket activation. Valid for "tcp" and "udp" sockets.
+
+
+</Field>
+
+
+<Field
+  common={false}
+  defaultValue={"host"}
+  enumValues={null}
+  examples={["host"]}
+  name={"host_key"}
+  nullable={false}
+  path={null}
+  relevantWhen={null}
+  required={false}
+  templateable={false}
+  type={"string"}
+  unit={null}
+  >
+
+### host_key
+
+The key name added to each event representing the current host. Valid for "tcp" and "udp" sockets. See [Context](#context) for more info.
+
+
+</Field>
+
+
+<Field
+  common={false}
+  defaultValue={102400}
+  enumValues={null}
+  examples={[102400]}
+  name={"max_length"}
+  nullable={true}
+  path={null}
+  relevantWhen={null}
+  required={false}
+  templateable={false}
+  type={"int"}
+  unit={"bytes"}
+  >
+
+### max_length
+
+The maximum bytes size of incoming messages before they are discarded. Valid for "tcp" and "udp" sockets.
+
+
+</Field>
+
+
+<Field
+  common={true}
+  defaultValue={null}
+  enumValues={{"tcp":"The TCP protocol.","udp":"The UDP protocol."}}
+  examples={["tcp","udp"]}
+  name={"mode"}
+  nullable={false}
+  path={null}
+  relevantWhen={null}
+  required={true}
+  templateable={false}
+  type={"string"}
+  unit={null}
+  >
+
+### mode
+
+The type of socket to use.
+
+
+</Field>
+
+
+<Field
+  common={false}
+  defaultValue={30}
+  enumValues={null}
+  examples={[30]}
+  name={"shutdown_timeout_secs"}
+  nullable={false}
+  path={null}
+  relevantWhen={null}
+  required={false}
+  templateable={false}
+  type={"int"}
+  unit={"seconds"}
+  >
+
+### shutdown_timeout_secs
+
+The timeout before a connection is forcefully closed during shutdown. Valid only for "tcp" sockets.
+
+
+</Field>
+
+
+</Fields>
+
+## Output
+
+The `socket` source ingests data through a socket, such as a TCP, UDP, or Unix socket and outputs [`log`][docs.data-model.log] events.
+For example:
+
+
+
+
+More detail on the output schema is below.
+
+<Fields filters={true}>
+
+
+<Field
+  defaultValue={null}
+  enumValues={null}
+  examples={["my.host.com"]}
+  name={"host"}
+  path={null}
+  required={true}
+  type={"string"}
+  >
+
+### host
+
+The upstream hostname.
+
+
+
+</Field>
+
+
+<Field
+  defaultValue={null}
+  enumValues={null}
+  examples={["Started GET / for 127.0.0.1 at 2012-03-10 14:28:14 +0100"]}
+  name={"message"}
+  path={null}
+  required={true}
+  type={"string"}
+  >
+
+### message
+
+The raw message, unaltered.
+
+
+
+</Field>
+
+
+<Field
+  defaultValue={null}
+  enumValues={null}
+  examples={["2019-11-01T21:15:47+00:00"]}
+  name={"timestamp"}
+  path={null}
+  required={true}
+  type={"timestamp"}
+  >
+
+### timestamp
+
+The exact time the event was ingested.
+
+
+</Field>
+
+
+</Fields>
+
+## How It Works
+
+### Context
+
+By default, the `socket` source will add context
+keys to your events via the [`host_key`](#host_key)
+options.
+
+### Environment Variables
+
+Environment variables are supported through all of Vector's configuration.
+Simply add `${MY_ENV_VAR}` in your Vector configuration file and the variable
+will be replaced before being evaluated.
+
+You can learn more in the [Environment Variables][docs.configuration#environment-variables]
+section.
+
+
+[docs.configuration#environment-variables]: /docs/setup/configuration/#environment-variables
+[docs.data-model.log]: /docs/about/data-model/log/
