--- website/docs/reference/sources/tcp.md.erb
@@ -1,34 +1,0 @@
-<%- component = metadata.sources.tcp -%>
-
-<%= component_header(component) %>
-
-## Configuration
-
-<%= component_config_example(component) %>
-
-## Options
-
-<%= options(component.specific_options_list, heading_depth: 3) %>
-
-<%- if component.env_vars_list.any? -%>
-## Env Vars
-
-<%= options(component.env_vars_list, heading_depth: 3) %>
-
-<%- end -%>
-<%- if component.output.log -%>
-## Output
-
-<%= component_output(component, component.output.log, heading_depth: 3) %>
-
-<%- end -%>
-## How It Works [[sort]]
-
-<%= component_sections(component) %>
-
-### Line Delimiters
-
-Each line is read until a new line delimiter (the `0xA` byte) is found.
-
-
-
