+++ website/metadata.js
@@ -18277,6 +18277,28 @@ module.exports = {
 
       ]
     },
+    "socket": {
+      "beta": false,
+      "delivery_guarantee": "best_effort",
+      "description": "Ingests data through a socket, such as a TCP, UDP, or Unix socket and outputs log events.",
+      "event_types": [
+        "log"
+      ],
+      "function_category": "receive",
+      "id": "socket_source",
+      "name": "socket",
+      "operating_systems": [
+        "linux",
+        "macos",
+        "windows"
+      ],
+      "service_provider": null,
+      "status": "prod-ready",
+      "type": "source",
+      "unsupported_operating_systems": [
+
+      ]
+    },
     "splunk_hec": {
       "beta": true,
       "delivery_guarantee": "at_least_once",
@@ -18353,50 +18375,6 @@ module.exports = {
       "function_category": "receive",
       "id": "syslog_source",
       "name": "syslog",
-      "operating_systems": [
-        "linux",
-        "macos",
-        "windows"
-      ],
-      "service_provider": null,
-      "status": "prod-ready",
-      "type": "source",
-      "unsupported_operating_systems": [
-
-      ]
-    },
-    "tcp": {
-      "beta": false,
-      "delivery_guarantee": "best_effort",
-      "description": "Ingests data through the TCP protocol and outputs log events.",
-      "event_types": [
-        "log"
-      ],
-      "function_category": "receive",
-      "id": "tcp_source",
-      "name": "tcp",
-      "operating_systems": [
-        "linux",
-        "macos",
-        "windows"
-      ],
-      "service_provider": null,
-      "status": "prod-ready",
-      "type": "source",
-      "unsupported_operating_systems": [
-
-      ]
-    },
-    "udp": {
-      "beta": false,
-      "delivery_guarantee": "best_effort",
-      "description": "Ingests data through the UDP protocol and outputs log events.",
-      "event_types": [
-        "log"
-      ],
-      "function_category": "receive",
-      "id": "udp_source",
-      "name": "udp",
       "operating_systems": [
         "linux",
         "macos",
